backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic>=2.7.0
pydantic-settings==2.9.1
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
transformers==4.35.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.0
pillow==10.1.0
pytesseract==0.3.10
langchain==0.0.335
redis==5.0.1
celery==5.3.4
langgraph==0.0.15
langchain-openai==0.0.2
langchain-community==0.0.13
langchain-core==0.1.15
aiofiles==23.2.1
aiohttp==3.9.1 